requirements.txt
streamlit
openai-whisper
yt-dlp
google-generativeai
fal_client
pillow
requests
